Construct a read or write connection into an archive file

Construct a read or write connection into an archive file

archive_read(archive, file = 1L, mode = "r", format = NULL,
  filter = NULL)

Arguments

archive
An archive object or character vector to the archive

file
The file to open the connection to. Can also be an numeric index into the `archive()` data.frame.

mode
The mode to open the file in.

format
`character(1)` default: `NULL` The archive format, one of ‘7zip’, ‘cab’, ‘cpio’, ‘iso9660’, ‘lha’, ‘mtree’, ‘shar’, ‘rar’, ‘tar’, ‘xar’, ‘zip’, ‘warc’ .

filter
`character(1)` default: `NULL` The archive filter, one of ‘none’, ‘gzip’, ‘bzip2’, ‘compress’, ‘lzma’, ‘xz’, ‘uuencode’, ‘lzip’, ‘lrzip’, ‘lzop’, ‘grzip’, ‘lz4’ .

Details

libarchive versions prior to 3.1.0 did not support explicit setting of the format and filter, instead relying on a bidding process to automatically determine the format of the archive. This automatic detection is also used when `format` or `filter` is `NULL`.

Examples

a <- archive(system.file(package = "archive", "extdata", "data.zip"))
# See files in archive
a

#> # A tibble: 3 x 3
#>             path  size                date
#>            <chr> <dbl>              <dttm>
#> 1       iris.csv   192 2017-04-28 15:55:29
#> 2     mtcars.csv   274 2017-04-28 15:55:29
#> 3 airquality.csv   142 2017-04-28 15:55:29

# By default reads the first file in the archive.
read.csv(archive_read(a), nrows = 3)

#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa

# Can also specify a filename directly
read.csv(archive_read(a, "mtcars.csv"), nrows = 3)

#>    mpg cyl disp  hp drat    wt  qsec vs am gear carb
#> 1 21.0   6  160 110 3.90 2.620 16.46  0  1    4    4
#> 2 21.0   6  160 110 3.90 2.875 17.02  0  1    4    4
#> 3 22.8   4  108  93 3.85 2.320 18.61  1  1    4    1

# Or by position
read.csv(archive_read(a, 3), nrows = 3)

#>   Ozone Solar.R Wind Temp Month Day
#> 1    41     190  7.4   67     5   1
#> 2    36     118  8.0   72     5   2
#> 3    12     149 12.6   74     5   3

# Explicitly specify the format and filter if automatic detection fails.
read.csv(archive_read(a, format = "zip"), nrows = 3)

#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa