Construct a write only connection to a new archive

`archive_read()` returns an readable input connection to an existing archive. `archive_write()` returns an writable output connection to a new archive.

archive_write(archive, file, format = NULL, filter = NULL)

archive_write_dir(archive, dir, ..., recursive = TRUE, full.names = FALSE)

Arguments

archive
`character(1)` The archive filename.

file
`character(1)` The filename within the archive.

format
`character(1)` default: `NULL` The archive format, one of ‘7zip’, ‘cab’, ‘cpio’, ‘iso9660’, ‘lha’, ‘mtree’, ‘shar’, ‘rar’, ‘tar’, ‘xar’, ‘zip’, ‘warc’ .

filter
`character(1)` default: `NULL` The archive filter, one of ‘none’, ‘gzip’, ‘bzip2’, ‘compress’, ‘lzma’, ‘xz’, ‘uuencode’, ‘lzip’, ‘lrzip’, ‘lzop’, ‘grzip’, ‘lz4’ .

dir
The directory of files to add

...
additional paramters passed to `base::dir`

recursive
logical. Should the listing recurse into directories?

full.names
a logical value. If TRUE, the directory path is prepended to the file names to give a relative file path. If FALSE, the file names (rather than paths) are returned.

Details

If `format` and `filter` are `NULL`, they will be set automatically based on the file extension given in `file`.

Examples

# Achive format and filters can be set automatically from the file extensions.
f1 <- tempfile(fileext = ".tar.gz")

write.csv(mtcars, archive_write(f1, "mtcars.csv"))
archive(f1)

#> # A tibble: 1 x 3
#>         path  size                date
#>        <chr> <dbl>              <dttm>
#> 1 mtcars.csv  1783 2017-07-27 13:06:50

unlink(f1)

# They can also be specified explicitly
f2 <- tempfile()
write.csv(iris, archive_write(f2, "iris.csv", format = "tar", filter = "bzip2"))
archive(f2)

#> # A tibble: 1 x 3
#>       path  size                date
#>      <chr> <dbl>              <dttm>
#> 1 iris.csv  4821 2017-07-27 13:06:50

unlink(f2)