Construct a connections for (possibly compressed) files.

These work similar to R's built-in connections for files and differ from archive_read and archive_write because they do not use an archive format, just use one or more of the filters.

file_write(file, filter = NULL)

file_read(file, mode = "r")

Arguments

file
character(1) || integer(1) The filename within the archive, specified either by filename or by position.

filter
character(1) default: NULL The archive filter, one of ‘none’, ‘gzip’, ‘bzip2’, ‘compress’, ‘lzma’, ‘xz’, ‘uuencode’, ‘lzip’, ‘lrzip’, ‘lzop’, ‘grzip’, ‘lz4’ .

mode
character(1) A description of how to open the connection (if it should be opened initially). See section ‘Modes’ in base::connections() for possible values.

Details

file_write() returns an writable output connection, file_read() returns a readable input connection.

Examples

if (archive:::libarchive_version() > "3.2.0") {
# Write bzip2, base 64 encoded data
write.csv(mtcars, file_write("mtcars.bz2", c("uuencode", "bzip2")))

# Read it back
read.csv(file_read("mtcars.bz2"), row.names = 1, nrows = 3)
unlink("mtcars.bz2")
}